moved files to model package
