+++ src/test/java/ithacacollege/comp345/group4/classPlanner/model/TranscriptTest.java
@@ -1,4 +1,4 @@
-package ithacacollege.comp345.group4.classPlanner;
+package ithacacollege.comp345.group4.classPlanner.model;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
